🍕 добавлен метод оплучение uuid
