+++ internal/utils/interface.go
@@ -1,0 +1,16 @@
+package utils
+
+//go:generate mockgen -destination=mock/mocks.go -source=interface.go
+
+type Interface interface {
+	EncryptPassword(publicKey *string, password string) (string, error)
+
+	UUID() string
+}
+
+type Utils struct {
+}
+
+func New() *Utils {
+	return &Utils{}
+}
